backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
opencv-python==4.9.0.80
grad-cam==1.5.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.0
